fix: gcc: implicit_casting

+++ libft/ft_printf/print_conversion/floating_point/utils/pf_calc_array.c
@@ -6,7 +6,7 @@
 /*   By: stanaka2 <stanaka2@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/06/03 13:46:30 by stanaka2          #+#    #+#             */
-/*   Updated: 2026/06/03 19:29:28 by stanaka2         ###   ########.fr       */
+/*   Updated: 2026/06/23 12:37:53 by stanaka2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ void	pf_array_half(uint8_t *dividend, int array_size)
 	i = 0;
 	while (i < array_size - 1)
 	{
-		dividend[i + 1] += (dividend[i] % 2) * 10;
+		dividend[i + 1] += (uint8_t)((dividend[i] % 2) * 10);
 		dividend[i] /= 2;
 		i++;
 	}
@@ -53,7 +53,7 @@ void	pf_array_double(uint8_t *multiplicand, int array_size)
 	carry = 0;
 	while (i >= 0)
 	{
-		multiplicand[i] = multiplicand[i] * 2 + carry;
+		multiplicand[i] = (uint8_t)(multiplicand[i] * 2 + carry);
 		carry = multiplicand[i] / 10;
 		multiplicand[i] %= 10;
 		i--;
@@ -69,7 +69,7 @@ void	pf_array_double_hex(uint8_t *multiplicand, int array_size)
 	carry = 0;
 	while (i >= 0)
 	{
-		multiplicand[i] = multiplicand[i] * 2 + carry;
+		multiplicand[i] = (uint8_t)(multiplicand[i] * 2 + carry);
 		carry = multiplicand[i] / 16;
 		multiplicand[i] %= 16;
 		i--;
