+++ libft/ft_strtod/utils/strtod_scan_digits.c
@@ -6,7 +6,7 @@
 /*   By: stanaka2 <stanaka2@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/06/14 03:22:31 by stanaka2          #+#    #+#             */
-/*   Updated: 2026/06/15 15:49:36 by stanaka2         ###   ########.fr       */
+/*   Updated: 2026/06/15 15:50:55 by stanaka2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -43,7 +43,7 @@ void	strtod_scan_decimal_digits(const char **nptr, t_to_double *to_double)
 			else if (shift < 0 && LONG_MAX + shift < i)
 				to_double->has_sticky = true;
 			else
-				add_digit(to_double, i - shift, **nptr - '0');
+				add_digit(to_double, i - shift, (uint8_t)(**nptr - '0'));
 		}
 		if (**nptr != '.')
 			++i;
@@ -55,7 +55,7 @@ void	strtod_scan_hex_digits(const char **nptr, t_to_double *to_double)
 {
 	const long		exponent = strtod_pre_scan_hex_exponent(*nptr);
 	const long		shift = exponent / 4;
-	const int8_t	mod = exponent % 4;
+	const int8_t	mod = (int8_t)(exponent % 4);
 	const char		*radix_point;
 	ptrdiff_t		i;
 
@@ -96,13 +96,13 @@ static void	add_scaled_hex_digits(\
 	{
 		add_digit(to_double, i, (n * power) % 16);
 		if (LONG_MIN < i)
-			add_digit(to_double, i - 1, (n * power) / 16);
+			add_digit(to_double, i - 1, (uint8_t)((n * power) / 16));
 	}
 	else if (mod < 0)
 	{
 		add_digit(to_double, i, n / power);
 		if (i < LONG_MAX)
-			add_digit(to_double, i + 1, ((n % power) * 16) / power);
+			add_digit(to_double, i + 1, (uint8_t)(((n % power) * 16) / power));
 	}
 	else
 		add_digit(to_double, i, n);
@@ -111,11 +111,11 @@ static void	add_scaled_hex_digits(\
 static uint8_t	hex_to_decimal(char c)
 {
 	if (ft_isupper(c))
-		return (c - 'A' + 10);
+		return ((uint8_t)(c - 'A' + 10));
 	else if (ft_islower(c))
-		return (c - 'a' + 10);
+		return ((uint8_t)(c - 'a' + 10));
 	else
-		return (c - '0');
+		return ((uint8_t)(c - '0'));
 }
 
 static void	add_digit(t_to_double *to_double, ptrdiff_t i, uint8_t digit)
